requirements.txt
# Performance Observer Suite - Python dependencies
#
# Core / always required
pyyaml>=6.0
requests>=2.28.0
numpy>=1.24.0
matplotlib>=3.7.0

# Kubernetes API client (required for K8s and Argo-via-kubeconfig collectors)
kubernetes>=28.1.0

# Optional: faster YAML parser
# pyyaml[libyaml]>=6.0
